requirements.txt
django
django-extentions
MySQL-python
biopython
nibabel

